Created four new functions -- allCourses, professorTotalGPA, calculateSectionGPA, and professorSections.
These functions include code to convert the raw grades into a GPA figure. These functions also reformat the data into a more professor-centric format, and make it easier to obtain certain values from the dataset. The allCourses() value returns a double array that groups all sections by professor, and includes a GPA key + value in the course object. The professorTotalGPA takes the professors name as a string and returns their average GPA across all semesters and sections for that particular class. The calculateSectionGPA takes in professors name, section, and semester, and returns the GPA for that particular section. The professorSections array takes in the professors name and outputs an array of javascript objects representing the sections they taught across all semesters.